requirements.txt
annotated-types
exceptiongroup
fastapi-pagination
pydantic
pydantic_core
PyJWT
python-decouple
python-dotenv
SQLAlchemy
typing_extensions
uvicorn==0.23.2
gunicorn
fastapi[all]
uvicorn[standard]
sqlalchemy
python-multipart
passlib
pydantic-settings
motor
pymongo
aiohttp
tenacity
pydantic-settings